Created new types for the display driver.
